requirements.txt
streamlit==1.31.1
mysql-connector-python==8.2.0